refactor(backend): rename `reviews` mod to `misc`
