+++ apps/backend/src/main.rs
@@ -47,8 +47,8 @@ mod entities;
 mod graphql;
 mod media;
 mod migrator;
+mod misc;
 mod movies;
-mod reviews;
 mod shows;
 mod users;
 mod utils;
